NodeEditor: fix font issue when element pop in
MOstly seen with cme meta edit. In node editor, edit an element, set the
type to node or leaf. The element appearing in the editor now have the
same font as the rest of the widget.

But the font propagation is a hack. The setup of -font with ConfigSpecs
does not work.