show feedback when pasting on wrong value type
pasting on node is not possible because we don't know where user's data
should land.

Likewise, pasting on a check list is difficult to act on. Does the user
want to select a pasted item ? Do unlisted items be deselected ?
From user's point of view, clicking on edit button and selecting items
is probably faster than playing with cut'n'paste.