TkUI: fix element filter
Before this fix, filtering 'User' in ssh model did not show the User
elements located in Host sections. This occured because the action of
the element was not stored in $data_ref->{actions}.

I've fixed another bug where not enough elements were hidden when a
filter is active.

Last but not least, with the current algorithm applying a filter and
ticking "show custom values" filters elements except if they have a
custom value, i.e. the custom value trumps the filter. I'm not sure
this is the correct behavior, but let's see if this is a problem in
real life before adding extra complexity to the filter...